Add static resource handler for uploaded images in FileConfiguration
This commit configures a resource handler for serving uploaded images statically from the filesystem in the FileConfiguration class. It maps the URL path '/uploaded-images/**' to the physical directory on the disk at 'file:///C:/Users/SHIN SUJUNG/IdeaProjects/ShoppingMall/uploaded-images/'. This allows images stored in this directory to be accessible via the specified path in the application.